requirements.txt
gevent>=1.0
pysnmp
pysnmp-mibs
pysmi
lxml
requests
sphinx
libtaxii>=1.1.0
crc16
natsort
scapy==2.4.3rc1
hpfeeds3
modbus-tk
stix-validator
stix
cybox
bacpypes==0.17.0
pyghmi==1.4.1
mixbox
modbus-tk
cpppo
fs==2.3.0
python-slugify
tftpy
# some freezegun versions broken
freezegun!=0.3.13
pytest
pycrypto
sphinx_rtd_theme
psutil
